==> 2.Data_Refinement.ipynb <==
import numpy as np
import pandas as pd
import re

## Refine The Data

Here we are mainly computing three three things :-
* Items
* Users 
* Interaction-Matrix

### Creation Of the Items

#### **Items** :- item_id + Metadata features

Now focussing on the items section :-

* movie_id
* title
* year(release)
* genre categories

Shifting to the item_feature dataframe:-
* overview
* language (original_language)
* runtime
* vote_average
* vote_count

### Coding  

items_raw=pd.read_csv("data/items_raw.csv")

item_features=pd.read_csv("data/item_features.csv")

1. we are extracting year feature from the release_title

items_raw["release_date"]=pd.to_datetime(items_raw.release_date,infer_datetime_format=True)
items_raw["year"]=items_raw.release_date.apply(lambda x: str(x.year))

2. Drop imdb_url,video_release_date & release_date

items_raw.columns

items_main=items_raw.drop(["IMDb_URL","video_release_date"],axis=1).copy()

items_main.head()

3. Get additional features from the the item_features

items_addtl = item_features[['overview', 'original_language', 'runtime', 'vote_average', 'vote_count', "movie_id"]].copy()

items_addtl

## Merging the two Dataframes

items = pd.merge(left=items_main, right=items_addtl, on="movie_id", how="left")

items.head()

items.to_csv("data/items.csv", index=None)

### Using the NLP libraries for Understanding The Function ##

import spacy

from spacy.lang.en.examples import sentences 

import spacy
from spacy.lang.en.examples import sentences 

nlp = spacy.load('en_core_web_lg')
doc = nlp(sentences[0])
print(doc.text)
for token in doc:
    print(token.text, token.pos_, token.dep_)

import spacy

#nlp = spacy.load('en_core_web_md')  # make sure to use larger model!
tokens = nlp(u'dog cat banana')

for token1 in tokens:
    for token2 in tokens:
        print(token1.text, token2.text, token1.similarity(token2))

import umap
from sklearn.datasets import load_digits

digits = load_digits()

embedding = umap.UMAP().fit_transform(digits.data)
