==> 3.Data_Transform.ipynb <==
## Transform

We need to do a number of operations on the data before running the algorithms. They are :-
* Data_Preprocessing
* Data_Splitting

#### Data_Preprocessing

* Users: Convert to label encoded values
* Items:  Convert to label encoded values
* User Side Features: Convert to
* Ratings: Convert the interaction to Explicit or Implicit signals


import sys
sys.path.append("Documents/Recommender System/")

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

users = pd.read_csv("data/users.csv")
items = pd.read_csv("data/items.csv")
ratings = pd.read_csv("data/ratings.csv")

users.head()

ratings.head()

items.columns

## Processing

from reco.preprocess import encode_user_item

ratings['rating'] = ratings['rating'].values.astype(np.float32)
min_rating = min(ratings['rating'])
max_rating = max(ratings['rating'])

from reco.preprocess import random_split

encoded_ratings, user_encoder, item_encoder = encode_user_item(ratings, "user_id", "movie_id","rating","unix_timestamp")

## Random Split

from reco.preprocess import random_split, user_split, sample_data

sampledf = sample_data()

train_random, val_random, test_random = random_split(sampledf, [0.6, 0.2, 0.2])

train_random

val_random

test_random

## User Chronological Split

def user_split(df,ratios,chrono=False):
    seed=42
    samples = df.shape[0]
    col_time="timestamp"
    col_user="user_index"
    splits=[]
    if chrono==True:
        df_grouped=df.sort_values(col_time).groupby(col_user)
    else:
        df_grouped=df.groupby(col_user)
    for name, group in df_grouped:
    
        group_splits = random_split(df_grouped.get_group(name), ratios, shuffle=False)
        
        # Concatenate the list of split dataframes.
        concat_group_splits = pd.concat(group_splits)
        splits.append(concat_group_splits)
    # Concatenate splits for all the groups together.
    splits_all = pd.concat(splits)

    # Take split by split_index
    splits_list = [ splits_all[splits_all["split_index"] == x] for x in range(len(ratios))]
    return splits_list


train_chrono, val_chrono, test_chrono = user_split(sampledf, [0.6, 0.2, 0.2])

train_chrono

val_chrono

test_chrono


ratings
